add coverage

+++ deeplake/core/meta/encode/chunk_id.py
@@ -285,11 +285,11 @@ class ChunkIdEncoder(Encoder, DeepLakeMemoryObject):
         if rows:
             for row in rows:
                 prev = (
-                    np.uint64(-1)
-                    if row == 0
-                    else self._encoded[row - 1][LAST_SEEN_INDEX_COLUMN]
+                    -1 if row == 0 else self._encoded[row - 1][LAST_SEEN_INDEX_COLUMN]
                 )
-                num_samples_in_chunk = self._encoded[row][LAST_SEEN_INDEX_COLUMN] - prev
+                num_samples_in_chunk = int(
+                    self._encoded[row][LAST_SEEN_INDEX_COLUMN] - prev
+                )
                 self._encoded[row:, LAST_SEEN_INDEX_COLUMN] -= num_samples_in_chunk
             self._encoded = np.delete(self._encoded, rows, axis=0)
             self.is_dirty = True
@@ -310,12 +310,8 @@ class ChunkIdEncoder(Encoder, DeepLakeMemoryObject):
             to_delete = self._delete_rows(rows)
         else:
             row = rows[0]
-            prev = (
-                np.uint64(-1)
-                if row == 0
-                else self._encoded[row - 1][LAST_SEEN_INDEX_COLUMN]
-            )
-            num_samples_in_chunk = self.array[row][LAST_SEEN_INDEX_COLUMN] - prev
+            prev = -1 if row == 0 else self._encoded[row - 1][LAST_SEEN_INDEX_COLUMN]
+            num_samples_in_chunk = int(self.array[row][LAST_SEEN_INDEX_COLUMN] - prev)
 
             if num_samples_in_chunk == 1:
                 to_delete = self._delete_rows(rows)
